+++ global/js/background.js
@@ -71,6 +71,7 @@ var Background = {
    **/
   init : function() {
     Background.checkIfUpdatedSilently();
+    sttb.contextMenus.toggle();
   }
   ,